requirements.txt
flask
pytorch
pillow
numpy
opencv